--- a/.github/workflows/benchmark.yml
+++ b/.github/workflows/benchmark.yml
@@ -11,6 +11,14 @@ on:
       - 'Dockerfile'
       - '.dockerignore'
       - '**.sample'
+  pull_request:
+    paths-ignore:
+      - '**.md'
+      - 'LICENSE'
+      - '**/.gitignore'
+      - 'Dockerfile'
+      - '.dockerignore'
+      - '**.sample'
 
 env:
   # See: Makefile