## !!! BREAKING CHANGES !!!

## Fix Security Problems

## Add New Features

## Improve Current Featues

- [x] fix excessive use of mutex lock, and improve performance

## Other

- [x] upgrade dependency modules

## Appendix

### benchmark

- benchmark of this branch is [here](https://github.com/goldeneggg/structil/actions/runs/2740904700) (see: artifact named `benchmark`)
- benchmark of v0.9.0 is [here](https://github.com/goldeneggg/structil/blob/a0b72ad701e401019e18f106d6061ed61a0085b9/BENCHMARK_LATEST.txt)
- compare benchmark result of this branch and v0.9.0 is as follows (by `benchstat`)
  - original compare artifact file is [here](https://github.com/goldeneggg/structil/actions/runs/2740904700) (see: artifact named `benchstat-vs-latest`)

```
name                                          old time/op    new time/op    delta
pkg:github.com/goldeneggg/structil goos:linux goarch:amd64
NewGetter_Val                                   7.65µs ± 4%    6.73µs ± 0%  -12.01%  (p=0.008 n=5+5)
NewGetter_Val-2                                 7.85µs ± 2%    6.99µs ± 8%  -10.96%  (p=0.008 n=5+5)
NewGetter_Ptr                                   8.91µs ± 1%    7.60µs ± 0%  -14.74%  (p=0.008 n=5+5)
NewGetter_Ptr-2                                 9.51µs ± 8%    7.58µs ± 0%  -20.30%  (p=0.008 n=5+5)
GetterGetType_String                            37.6ns ± 1%    12.6ns ± 4%  -66.35%  (p=0.008 n=5+5)
GetterGetType_String-2                          37.0ns ± 3%    13.0ns ± 5%  -64.80%  (p=0.008 n=5+5)
GetterGetValue_String                           36.6ns ± 2%    12.9ns ± 2%  -64.83%  (p=0.008 n=5+5)
GetterGetValue_String-2                         36.9ns ± 2%    13.1ns ± 5%  -64.50%  (p=0.008 n=5+5)
GetterHas_String                                36.7ns ± 2%    12.4ns ± 0%  -66.29%  (p=0.016 n=5+4)
GetterHas_String-2                              36.9ns ± 2%    12.6ns ± 3%  -65.73%  (p=0.008 n=5+5)
GetterGet_String                                36.6ns ± 1%    13.0ns ± 3%  -64.59%  (p=0.008 n=5+5)
GetterGet_String-2                              36.2ns ± 2%    13.1ns ± 3%  -63.85%  (p=0.008 n=5+5)
GetterString                                    41.2ns ± 2%    13.2ns ± 2%  -67.97%  (p=0.008 n=5+5)
GetterString-2                                  41.1ns ± 1%    13.2ns ± 3%  -67.87%  (p=0.008 n=5+5)
GetterBytes                                     42.1ns ± 1%    13.4ns ± 4%  -68.02%  (p=0.008 n=5+5)
GetterBytes-2                                   41.8ns ± 2%    13.1ns ± 2%  -68.54%  (p=0.008 n=5+5)
GetterUintptr                                   42.2ns ± 4%    14.5ns ± 7%  -65.68%  (p=0.008 n=5+5)
GetterUintptr-2                                 42.6ns ± 1%    14.4ns ± 8%  -66.26%  (p=0.008 n=5+5)
GetterUnsafePointer                             42.9ns ± 5%    13.7ns ± 5%  -68.12%  (p=0.008 n=5+5)
GetterUnsafePointer-2                           42.8ns ± 2%    14.9ns ± 6%  -65.14%  (p=0.008 n=5+5)
GetterSlice_StructPtrSlice                       140ns ± 2%      95ns ± 2%  -32.00%  (p=0.008 n=5+5)
GetterSlice_StructPtrSlice-2                     145ns ± 1%      96ns ± 2%  -33.91%  (p=0.008 n=5+5)
GetterGetGetter                                  831ns ± 3%     685ns ± 2%  -17.55%  (p=0.008 n=5+5)
GetterGetGetter-2                                859ns ± 1%     681ns ± 1%  -20.70%  (p=0.008 n=5+5)
GetterToMap                                     5.26µs ± 2%    4.66µs ± 0%  -11.50%  (p=0.008 n=5+5)
GetterToMap-2                                   5.25µs ± 3%    4.58µs ± 1%  -12.79%  (p=0.008 n=5+5)
GetterIsStruct                                  42.9ns ± 3%    16.8ns ±12%  -60.95%  (p=0.008 n=5+5)
GetterIsStruct-2                                42.8ns ± 9%    17.3ns ±35%  -59.61%  (p=0.008 n=5+5)
GetterIsSlice_Bytes                             40.4ns ± 6%    18.1ns ± 0%  -55.27%  (p=0.008 n=5+5)
GetterIsSlice_Bytes-2                           41.1ns ± 7%    16.4ns ±20%  -60.20%  (p=0.008 n=5+5)
GetterIsSlice_StructSlice                       43.8ns ± 8%    16.0ns ± 8%  -63.39%  (p=0.008 n=5+5)
GetterIsSlice_StructSlice-2                     42.3ns ± 4%    17.5ns ±11%  -58.55%  (p=0.008 n=5+5)
GetterIsSlice_StructPtrSlice                    42.7ns ± 3%    16.7ns ± 7%  -60.94%  (p=0.008 n=5+5)
GetterIsSlice_StructPtrSlice-2                  44.6ns ± 6%    19.8ns ±27%  -55.62%  (p=0.008 n=5+5)
GetterMapGet                                    2.55µs ± 2%    2.13µs ± 0%  -16.47%  (p=0.008 n=5+5)
GetterMapGet-2                                  2.58µs ± 1%    2.12µs ± 0%  -17.77%  (p=0.008 n=5+5)
NewFinder_Val                                   5.72µs ± 2%    4.94µs ± 4%  -13.58%  (p=0.008 n=5+5)
NewFinder_Val-2                                 5.91µs ± 1%    4.86µs ± 0%  -17.83%  (p=0.008 n=5+5)
NewFinder_Ptr                                   6.20µs ± 1%    5.47µs ± 0%  -11.77%  (p=0.008 n=5+5)
NewFinder_Ptr-2                                 6.36µs ± 3%    5.31µs ± 1%  -16.49%  (p=0.008 n=5+5)
ToMap_1FindOnly                                  461ns ± 1%     352ns ± 1%  -23.62%  (p=0.008 n=5+5)
ToMap_1FindOnly-2                                466ns ± 1%     362ns ± 1%  -22.32%  (p=0.008 n=5+5)
ToMap_2FindOnly                                  607ns ± 2%     449ns ± 1%  -26.01%  (p=0.008 n=5+5)
ToMap_2FindOnly-2                                605ns ± 1%     451ns ± 1%  -25.47%  (p=0.008 n=5+5)
ToMap_1Struct_1Find                              595ns ± 1%     490ns ± 0%  -17.70%  (p=0.008 n=5+5)
ToMap_1Struct_1Find-2                            612ns ± 1%     489ns ± 1%  -20.15%  (p=0.008 n=5+5)
ToMap_1Struct_1Find_2Pair                        956ns ± 0%     743ns ± 0%  -22.35%  (p=0.008 n=5+5)
ToMap_1Struct_1Find_2Pair-2                      959ns ± 1%     738ns ± 1%  -23.04%  (p=0.008 n=5+5)
ToMap_2Struct_1Find                              800ns ± 1%     651ns ± 0%  -18.56%  (p=0.008 n=5+5)
ToMap_2Struct_1Find-2                            792ns ± 1%     648ns ± 2%  -18.24%  (p=0.008 n=5+5)
ToMap_2Struct_2Find                             1.00µs ± 1%    0.79µs ± 0%  -21.29%  (p=0.008 n=5+5)
ToMap_2Struct_2Find-2                           1.01µs ± 2%    0.86µs ±19%     ~     (p=0.103 n=5+5)
NewFinderKeys_yml                                143µs ± 3%     115µs ± 0%  -19.25%  (p=0.008 n=5+5)
NewFinderKeys_yml-2                              144µs ± 2%     114µs ± 0%  -20.82%  (p=0.008 n=5+5)
NewFinderKeys_json                              96.7µs ± 5%    76.1µs ± 0%  -21.25%  (p=0.008 n=5+5)
NewFinderKeys_json-2                            98.9µs ± 2%    75.7µs ± 0%  -23.47%  (p=0.008 n=5+5)
pkg:github.com/goldeneggg/structil/dynamicstruct goos:linux goarch:amd64
AddString                                        324ns ± 1%      76ns ± 0%  -76.43%  (p=0.008 n=5+5)
AddString-2                                      339ns ± 1%      78ns ± 7%  -77.07%  (p=0.008 n=5+5)
AddStringWithTag                                 328ns ± 1%      77ns ± 0%  -76.58%  (p=0.008 n=5+5)
AddStringWithTag-2                               339ns ± 2%      77ns ± 0%  -77.22%  (p=0.008 n=5+5)
AddInt                                           330ns ± 1%      77ns ± 0%  -76.71%  (p=0.008 n=5+5)
AddInt-2                                         342ns ± 1%      77ns ± 1%  -77.53%  (p=0.008 n=5+5)
AddFloa32                                        330ns ± 1%      77ns ± 1%  -76.54%  (p=0.008 n=5+5)
AddFloa32-2                                      343ns ± 2%      78ns ± 1%  -77.27%  (p=0.008 n=5+5)
AddFloa64                                        331ns ± 1%      78ns ± 0%  -76.42%  (p=0.008 n=5+5)
AddFloa64-2                                      347ns ± 2%      79ns ± 4%  -77.16%  (p=0.008 n=5+5)
AddBool                                          332ns ± 2%      78ns ± 0%  -76.62%  (p=0.008 n=5+5)
AddBool-2                                        338ns ± 3%      78ns ± 0%  -76.90%  (p=0.008 n=5+5)
AddMap                                           390ns ± 1%     130ns ± 0%  -66.74%  (p=0.008 n=5+5)
AddMap-2                                         422ns ±13%     130ns ± 0%  -69.16%  (p=0.008 n=5+5)
AddFunc                                          782ns ± 2%     452ns ± 0%  -42.21%  (p=0.008 n=5+5)
AddFunc-2                                        796ns ± 0%     451ns ± 0%  -43.36%  (p=0.016 n=4+5)
AddChanBoth                                      385ns ± 2%     119ns ± 1%  -69.03%  (p=0.008 n=5+5)
AddChanBoth-2                                    394ns ± 1%     120ns ± 0%  -69.54%  (p=0.008 n=5+5)
AddChanRecv                                      381ns ± 1%     122ns ± 0%  -67.98%  (p=0.008 n=5+5)
AddChanRecv-2                                    399ns ± 2%     121ns ± 0%  -69.59%  (p=0.008 n=5+5)
AddChanSend                                      386ns ± 1%     123ns ± 0%  -68.05%  (p=0.008 n=5+5)
AddChanSend-2                                    399ns ± 1%     123ns ± 0%  -69.16%  (p=0.008 n=5+5)
AddStruct                                       11.3µs ± 1%     9.3µs ± 0%  -17.42%  (p=0.008 n=5+5)
AddStruct-2                                     11.5µs ± 1%     9.3µs ± 0%  -19.17%  (p=0.008 n=5+5)
AddStructPtr                                    11.0µs ± 1%     9.3µs ± 0%  -15.44%  (p=0.008 n=5+5)
AddStructPtr-2                                  11.3µs ± 2%     9.2µs ± 0%  -19.08%  (p=0.008 n=5+5)
AddSlice                                         395ns ± 1%     129ns ± 0%  -67.32%  (p=0.008 n=5+5)
AddSlice-2                                       404ns ± 3%     129ns ± 1%  -68.13%  (p=0.008 n=5+5)
AddInterface                                     332ns ± 1%      80ns ± 0%  -75.83%  (p=0.008 n=5+5)
AddInterface-2                                   340ns ± 2%      81ns ± 1%  -76.26%  (p=0.008 n=5+5)
AddInterfacePtr                                  342ns ± 1%      94ns ± 0%  -72.70%  (p=0.008 n=5+5)
AddInterfacePtr-2                                350ns ± 2%      93ns ± 0%  -73.42%  (p=0.008 n=5+5)
Build                                           45.4µs ±10%    37.6µs ± 6%  -17.18%  (p=0.008 n=5+5)
Build-2                                         41.5µs ± 2%    32.2µs ± 3%  -22.50%  (p=0.008 n=5+5)
BuildNonPtr                                     51.1µs ± 3%    38.5µs ± 1%  -24.77%  (p=0.008 n=5+5)
BuildNonPtr-2                                   42.2µs ± 1%    32.2µs ± 4%  -23.75%  (p=0.008 n=5+5)
NewInterface                                     508ns ± 5%     366ns ± 1%  -27.95%  (p=0.008 n=5+5)
NewInterface-2                                   369ns ± 2%     268ns ± 4%  -27.55%  (p=0.008 n=5+5)
Definition                                      2.75ns ± 1%    2.47ns ± 0%  -10.40%  (p=0.016 n=5+4)
Definition-2                                    2.72ns ± 2%    2.47ns ± 0%   -9.25%  (p=0.008 n=5+5)
pkg:github.com/goldeneggg/structil/dynamicstruct/decoder goos:linux goarch:amd64
FromJSON_singleJSON                             8.67µs ± 2%    7.32µs ± 0%  -15.61%  (p=0.016 n=5+4)
FromJSON_singleJSON-2                           9.03µs ± 2%    7.32µs ± 2%  -18.99%  (p=0.008 n=5+5)
DynamicStruct_singleJSON_nonNest_nonUseTag      8.36µs ± 1%    6.79µs ± 0%  -18.84%  (p=0.008 n=5+5)
DynamicStruct_singleJSON_nonNest_nonUseTag-2    8.54µs ± 0%    6.69µs ± 0%  -21.62%  (p=0.008 n=5+5)
DynamicStruct_singleJSON_nest_nonUseTag         21.8µs ± 1%    18.1µs ± 0%  -17.04%  (p=0.008 n=5+5)
DynamicStruct_singleJSON_nest_nonUseTag-2       22.0µs ± 1%    17.8µs ± 0%  -19.10%  (p=0.008 n=5+5)
DynamicStruct_singleJSON_nest_useTag            36.9µs ± 1%    31.3µs ± 0%  -15.15%  (p=0.008 n=5+5)
DynamicStruct_singleJSON_nest_useTag-2          36.8µs ± 2%    30.9µs ± 0%  -15.83%  (p=0.008 n=5+5)
FromJSON_arrayJSON                              18.0µs ± 3%    15.8µs ± 0%  -12.10%  (p=0.008 n=5+5)
FromJSON_arrayJSON-2                            18.4µs ± 2%    15.8µs ± 1%  -14.46%  (p=0.008 n=5+5)
DynamicStruct_arrayJSON_nonNest_nonUseTag       8.00µs ± 1%    6.86µs ± 1%  -14.17%  (p=0.008 n=5+5)
DynamicStruct_arrayJSON_nonNest_nonUseTag-2     8.16µs ± 3%    6.70µs ± 0%  -17.89%  (p=0.008 n=5+5)
DynamicStruct_arrayJSON_nest_nonUseTag          20.9µs ± 2%    18.2µs ± 0%  -12.88%  (p=0.008 n=5+5)
DynamicStruct_arrayJSON_nest_nonUseTag-2        20.8µs ± 2%    17.9µs ± 1%  -13.76%  (p=0.008 n=5+5)
DynamicStruct_arrayJSON_nest_useTag             34.8µs ± 2%    31.4µs ± 0%   -9.59%  (p=0.008 n=5+5)
DynamicStruct_arrayJSON_nest_useTag-2           35.3µs ± 2%    31.2µs ± 1%  -11.63%  (p=0.008 n=5+5)
FromYAML_singleYAML                             53.2µs ± 3%    47.8µs ± 0%  -10.17%  (p=0.008 n=5+5)
FromYAML_singleYAML-2                           54.0µs ± 2%    46.8µs ± 0%  -13.26%  (p=0.008 n=5+5)
DynamicStruct_singleYAML_nonNest_nonUseTag      8.11µs ± 3%    6.84µs ± 0%  -15.68%  (p=0.008 n=5+5)
DynamicStruct_singleYAML_nonNest_nonUseTag-2    8.30µs ± 3%    6.70µs ± 0%  -19.28%  (p=0.008 n=5+5)
DynamicStruct_singleYAML_nest_nonUseTag         20.8µs ± 5%    18.2µs ± 0%  -12.55%  (p=0.008 n=5+5)
DynamicStruct_singleYAML_nest_nonUseTag-2       20.9µs ± 2%    18.0µs ± 1%  -14.14%  (p=0.008 n=5+5)
DynamicStruct_singleYAML_nest_useTag            20.7µs ± 2%    18.3µs ± 0%  -11.48%  (p=0.008 n=5+5)
DynamicStruct_singleYAML_nest_useTag-2          20.6µs ± 2%    17.9µs ± 0%  -13.47%  (p=0.008 n=5+5)
FromYAML_arrayYAML                               107µs ± 1%      95µs ± 0%  -10.75%  (p=0.008 n=5+5)
FromYAML_arrayYAML-2                             106µs ± 4%      93µs ± 0%  -12.29%  (p=0.008 n=5+5)
DynamicStruct_arrayYAML_nonNest_nonUseTag       7.92µs ± 3%    6.90µs ± 1%  -12.78%  (p=0.008 n=5+5)
DynamicStruct_arrayYAML_nonNest_nonUseTag-2     8.02µs ± 4%    6.73µs ± 0%  -16.15%  (p=0.008 n=5+5)
DynamicStruct_arrayYAML_nest_nonUseTag          20.8µs ± 3%    18.5µs ± 3%  -10.98%  (p=0.008 n=5+5)
DynamicStruct_arrayYAML_nest_nonUseTag-2        21.3µs ± 3%    17.9µs ± 0%  -16.22%  (p=0.008 n=5+5)
DynamicStruct_arrayYAML_nest_useTag             35.4µs ± 3%    31.6µs ± 0%  -10.65%  (p=0.008 n=5+5)
DynamicStruct_arrayYAML_nest_useTag-2           36.0µs ± 7%    31.1µs ± 0%  -13.73%  (p=0.008 n=5+5)
JSONToGetter_singleJSON_nonNest                 50.3µs ± 1%    43.3µs ± 0%  -13.88%  (p=0.008 n=5+5)
JSONToGetter_singleJSON_nonNest-2               50.4µs ± 2%    42.3µs ± 1%  -15.99%  (p=0.008 n=5+5)
JSONToGetter_singleJSON_nest                    71.7µs ± 1%    60.6µs ± 0%  -15.47%  (p=0.008 n=5+5)
JSONToGetter_singleJSON_nest-2                  70.5µs ± 2%    59.1µs ± 0%  -16.06%  (p=0.008 n=5+5)
JSONToGetter_arrayJSON_nonNest                  62.1µs ± 4%    52.5µs ± 0%  -15.46%  (p=0.016 n=5+4)
JSONToGetter_arrayJSON_nonNest-2                60.3µs ± 2%    51.0µs ± 1%  -15.47%  (p=0.008 n=5+5)
JSONToGetter_arrayJSON_nest                     80.8µs ± 2%    70.1µs ± 0%  -13.27%  (p=0.008 n=5+5)
JSONToGetter_arrayJSON_nest-2                   78.8µs ± 1%    68.0µs ± 0%  -13.70%  (p=0.008 n=5+5)
YAMLToGetter_singleYAML_nonNest                  180µs ± 2%     162µs ± 0%   -9.80%  (p=0.008 n=5+5)
YAMLToGetter_singleYAML_nonNest-2                179µs ± 2%     155µs ± 1%  -13.46%  (p=0.008 n=5+5)
YAMLToGetter_singleYAML_nest                     208µs ± 2%     183µs ± 0%  -11.98%  (p=0.008 n=5+5)
YAMLToGetter_singleYAML_nest-2                   205µs ± 2%     176µs ± 1%  -14.18%  (p=0.008 n=5+5)
YAMLToGetter_arrayYAML_nonNest                   241µs ± 3%     213µs ± 0%  -11.79%  (p=0.016 n=5+4)
YAMLToGetter_arrayYAML_nonNest-2                 242µs ± 4%     204µs ± 0%  -15.55%  (p=0.008 n=5+5)
YAMLToGetter_arrayYAML_nest                      270µs ± 2%     232µs ± 0%  -14.00%  (p=0.008 n=5+5)
YAMLToGetter_arrayYAML_nest-2                    270µs ± 2%     223µs ± 0%  -17.55%  (p=0.008 n=5+5)
pkg:github.com/goldeneggg/structil/util goos:linux goarch:amd64
ToI_String                                      8.60ns ± 1%    6.52ns ± 1%  -24.23%  (p=0.008 n=5+5)
ToI_String-2                                    8.57ns ± 1%    6.56ns ± 3%  -23.47%  (p=0.008 n=5+5)
ToI_Int                                         8.71ns ± 2%    6.51ns ± 0%  -25.31%  (p=0.016 n=5+4)
ToI_Int-2                                       8.65ns ± 1%    6.51ns ± 0%  -24.74%  (p=0.008 n=5+5)
ToI_StructPtr                                   8.51ns ± 1%    6.45ns ± 0%  -24.21%  (p=0.008 n=5+5)
ToI_StructPtr-2                                 8.56ns ± 1%    7.30ns ±16%  -14.76%  (p=0.008 n=5+5)
ToI_Map                                         8.50ns ± 2%    8.43ns ± 0%     ~     (p=0.635 n=5+5)
ToI_Map-2                                       8.51ns ± 1%    7.25ns ±16%  -14.83%  (p=0.032 n=5+5)
ToI_Func                                        8.45ns ± 4%    6.45ns ± 0%  -23.71%  (p=0.008 n=5+5)
ToI_Func-2                                      8.31ns ± 3%    6.45ns ± 0%  -22.42%  (p=0.008 n=5+5)
ToI_Chan                                        8.37ns ± 0%    6.51ns ± 4%  -22.15%  (p=0.008 n=5+5)
ToI_Chan-2                                      8.24ns ± 2%    6.55ns ± 2%  -20.42%  (p=0.008 n=5+5)
ToI_Nil                                         2.48ns ± 2%    2.25ns ± 4%   -9.50%  (p=0.008 n=5+5)
ToI_Nil-2                                       2.52ns ± 2%    2.25ns ± 1%  -10.81%  (p=0.008 n=5+5)
ElemTypeOf_String                               21.8ns ± 3%    18.8ns ± 0%  -13.75%  (p=0.008 n=5+5)
ElemTypeOf_String-2                             22.6ns ± 3%    18.7ns ± 0%  -16.90%  (p=0.008 n=5+5)
ElemTypeOf_Int                                  21.3ns ± 3%    18.8ns ± 0%  -11.85%  (p=0.008 n=5+5)
ElemTypeOf_Int-2                                22.3ns ± 3%    18.8ns ± 1%  -15.58%  (p=0.008 n=5+5)
ElemTypeOf_StructPtr                            21.4ns ± 3%    18.8ns ± 0%  -12.02%  (p=0.008 n=5+5)
ElemTypeOf_StructPtr-2                          22.4ns ± 3%    18.8ns ± 1%  -16.20%  (p=0.008 n=5+5)
ElemTypeOf_Map                                  21.7ns ± 3%    18.8ns ± 0%  -13.39%  (p=0.008 n=5+5)
ElemTypeOf_Map-2                                22.8ns ± 2%    18.8ns ± 0%  -17.70%  (p=0.008 n=5+5)
ElemTypeOf_Func                                 21.7ns ± 1%    18.8ns ± 0%  -13.49%  (p=0.008 n=5+5)
ElemTypeOf_Func-2                               22.8ns ± 1%    18.8ns ± 0%  -17.49%  (p=0.008 n=5+5)
ElemTypeOf_Chan                                 22.0ns ± 2%    18.8ns ± 0%  -14.39%  (p=0.008 n=5+5)
ElemTypeOf_Chan-2                               22.8ns ± 2%    18.8ns ± 0%  -17.50%  (p=0.008 n=5+5)
ElemTypeOf_Error                                21.8ns ± 2%    18.8ns ± 0%  -13.49%  (p=0.008 n=5+5)
ElemTypeOf_Error-2                              22.4ns ± 4%    18.7ns ± 0%  -16.26%  (p=0.008 n=5+5)
ElemTypeOf_Nil                                  21.0ns ± 3%    18.8ns ± 0%  -10.43%  (p=0.008 n=5+5)
ElemTypeOf_Nil-2                                22.5ns ± 2%    18.8ns ± 1%  -16.29%  (p=0.008 n=5+5)
RecoverToError                                  19.7µs ± 5%    17.2µs ± 0%  -12.60%  (p=0.008 n=5+5)
RecoverToError-2                                20.8µs ± 1%    17.3µs ± 0%  -16.95%  (p=0.008 n=5+5)

name                                          old alloc/op   new alloc/op   delta
pkg:github.com/goldeneggg/structil goos:linux goarch:amd64
NewGetter_Val                                   8.56kB ± 0%    8.53kB ± 0%   -0.37%  (p=0.029 n=4+4)
NewGetter_Val-2                                 8.56kB ± 0%    8.53kB ± 0%   -0.37%  (p=0.000 n=5+4)
NewGetter_Ptr                                   8.52kB ± 0%    8.49kB ± 0%   -0.37%  (p=0.000 n=4+5)
NewGetter_Ptr-2                                 8.52kB ± 0%    8.49kB ± 0%   -0.38%  (p=0.008 n=5+5)
GetterGetType_String                             0.00B          0.00B          ~     (all equal)
GetterGetType_String-2                           0.00B          0.00B          ~     (all equal)
GetterGetValue_String                            0.00B          0.00B          ~     (all equal)
…